Fix response caching bug...
The response cache stored a hash for the current station response
indefinitely. It then uses that hash value to compare it against the
current hash response, to determine for how long to TTL the current
response.

The idea is as follows: if the current hash is the same as the previous
hash, then the response (current playing track) didn't change, and thus
we want a short TTL to detect a track change. If the response did just
change though, we had a track change so we can affort a longer TTL.

The problem is this doesn't work if we store the response hash
indefinitely. If we fetch info from a domain at T and cache the response
hash, then query it again at T+30mins, the response hashes will be
different, but that doesn't say there was a track change recently, it
just says there was a track change during the last 30 mins... which
isn't surprising. The track could then change 2 seconds after we fetch
the info.

Address this by adding a ttl to the response-hash. We cache the
response-hash for as long as we'd cache the response if we think we
detected a track change.